Waggle - separate unit tests that use tor - they take a long time and are less stable
