+++ packages/waggle/package.json
@@ -14,7 +14,8 @@
 		"postversion": "git push && git push --tags",
 		"lint": "eslint --ext .jsx,.js,.ts,.tsx ./src/",
 		"test": "cross-env DEBUG=ipfs:*,waggle:* jest ./src/**/* --runInBand --verbose --testPathIgnorePatterns=\".src/(!?nodeTest*)|(.node_modules*)\"",
-		"test-ci": "jest ./src/**/* --runInBand --colors --ci --silent --verbose --testPathIgnorePatterns=\".src/(!?nodeTest*)|(.node_modules*)\"",
+		"test-ci": "DEBUG=waggle:* jest ./src/**/* --runInBand --colors --ci --silent --verbose --testPathIgnorePatterns=\".src/(!?nodeTest*)|(.node_modules*)|src/.*\\.tor.test\\.(t|j)s$\"",
+		"test-ci-tor": "DEBUG=waggle:* jest --runInBand --colors --ci --silent --verbose --testPathPattern=\"src/.*\\.tor.test\\.(t|j)s$\"",
 		"test-connect": "DEBUG='libp2p:websockets*' jest ./src/nodeTest/* --verbose",
 		"test-connect-ci": "jest ./src/nodeTest/* --colors --ci --silent --verbose",
 		"test-replication-no-tor": "ts-node -v && cross-env DEBUG='waggle:dbSnap*,waggle:localTest*' ts-node src/nodeTest/testReplicate.ts --nodesCount 1 --timeThreshold 200 --entriesCount 1000 --no-useTor",
