+++ packages/waggle/src/libp2p/connectionsManager.test.ts
@@ -2,7 +2,6 @@ import PeerId from 'peer-id'
 import { createTmpDir, TmpDir, tmpZbayDirPath } from '../common/testUtils'
 import * as utils from '../common/utils'
 import { ConnectionsManager } from './connectionsManager'
-jest.setTimeout(100_000)
 
 let tmpDir: TmpDir
 let tmpAppDataPath: string
@@ -16,25 +15,6 @@ beforeEach(() => {
 })
 
 describe('Connections manager', () => {
-  it('runs tor by default', async () => {
-    const ports = await utils.getPorts()
-    connectionsManager = new ConnectionsManager({
-      agentHost: 'localhost',
-      agentPort: ports.socksPort,
-      httpTunnelPort: ports.httpTunnelPort,
-      io: new utils.DummyIOServer(),
-      options: {
-        env: {
-          appDataPath: tmpAppDataPath
-        },
-        torControlPort: ports.controlPort
-      }
-    })
-    await connectionsManager.init()
-    expect(connectionsManager.tor.process).not.toBeNull()
-    await connectionsManager.tor.kill()
-  })
-
   it('inits only tor control if spawnTor is set to false', async () => {
     const torPassword = 'testTorPassword'
     const ports = await utils.getPorts()
@@ -57,31 +37,6 @@ describe('Connections manager', () => {
     const torControl = connectionsManager.tor.torControl
     expect(torControl.password).toEqual(torPassword)
     expect(torControl.params.port).toEqual(ports.controlPort)
-  })
-
-  it('creates network', async () => {
-    const ports = await utils.getPorts()
-    connectionsManager = new ConnectionsManager({
-      agentHost: 'localhost',
-      agentPort: ports.socksPort,
-      httpTunnelPort: ports.httpTunnelPort,
-      io: new utils.DummyIOServer(),
-      options: {
-        env: {
-          appDataPath: tmpAppDataPath
-        },
-        torControlPort: ports.controlPort
-      }
-    })
-    await connectionsManager.init()
-    const spyOnDestroyHiddenService = jest.spyOn(connectionsManager.tor, 'destroyHiddenService')
-    const network = await connectionsManager.createNetwork()
-    expect(network.hiddenService.onionAddress.split('.')[0]).toHaveLength(56)
-    expect(network.hiddenService.privateKey).toHaveLength(99)
-    const peerId = await PeerId.createFromJSON(network.peerId)
-    expect(PeerId.isPeerId(peerId)).toBeTruthy()
-    expect(await spyOnDestroyHiddenService.mock.results[0].value).toBeTruthy()
-    await connectionsManager.tor.kill()
   })
 
   it('throws error when tries to send certification request to the offline registrar', async () => {
