+++ packages/waggle/src/socket/IOProxy.test.ts
@@ -1,9 +1,7 @@
-import { ConnectionsManager } from '../libp2p/connectionsManager'
-import { createCertificatesTestHelper } from '../libp2p/tests/client-server'
-import { CertificateRegistration } from '../registration'
-import { Storage } from '../storage'
 import { createMinConnectionManager, createTmpDir, ResponseMock, tmpZbayDirPath, TorMock } from '../common/testUtils'
 import { getPorts } from '../common/utils'
+import { ConnectionsManager } from '../libp2p/connectionsManager'
+import { createCertificatesTestHelper } from '../libp2p/tests/client-server'
 import { EventTypesResponse } from './constantsReponse'
 import IOProxy from './IOProxy'
 
@@ -101,37 +99,4 @@ describe('IO proxy', () => {
     expect(observedIO).toBeCalledTimes(1)
     expect(observedIO).toBeCalledWith(EventTypesResponse.SEND_USER_CERTIFICATE, { id: 'someCommunityId', payload: registrarResponse })
   })
-})
-
-test('IO proxy closes all services', async () => {
-  const pems = await createCertificatesTestHelper('adres1.onion', 'adres2.onion')
-  const certs1 = {
-    cert: pems.servCert,
-    key: pems.servKey,
-    ca: [pems.ca]
-  }
-  const certs2 = {
-    cert: pems.userCert,
-    key: pems.userKey,
-    ca: [pems.ca]
-  }
-  const appDataPath = createTmpDir()
-  const ports = await getPorts()
-  const manager = createMinConnectionManager({
-    env: { appDataPath: tmpZbayDirPath(appDataPath.name) },
-    torControlPort: ports.controlPort
-  })
-  await manager.init()
-  const ioProxy = new IOProxy(manager)
-  await ioProxy.createCommunity('myCommunity1', certs1, pems.ca, pems.ca_key)
-  await ioProxy.createCommunity('myCommunity2', certs2)
-  const spyTorKill = jest.spyOn(manager.tor, 'kill')
-  const spyOnIo = jest.spyOn(manager.io, 'close')
-  const spyOnRegistrarStop = jest.spyOn(CertificateRegistration.prototype, 'stop')
-  const spyOnStorageStop = jest.spyOn(Storage.prototype, 'stopOrbitDb')
-  await ioProxy.closeAll()
-  expect(spyTorKill).toBeCalledTimes(1)
-  expect(spyOnRegistrarStop).toBeCalledTimes(1)
-  expect(spyOnStorageStop).toBeCalledTimes(2)
-  expect(spyOnIo).toBeCalledTimes(1)
 })
